Add RequireByIdAsync<T> method to RepositoryExtensions
A new extension method, `RequireByIdAsync<T>`, was added to the
`RepositoryExtensions` class. This method retrieves an entity
by its ID asynchronously and throws an `ArgumentNullException`
if the entity is not found.

The method includes an optional `CancellationToken` parameter
(defaulting to `default`) to support cancellation.

A `#pragma warning disable CA1068` directive was added to
suppress the warning about the `CancellationToken` parameter
not being the last parameter, followed by a corresponding
`#pragma warning restore CA1068` directive.